[CHANGE] Refactor CSS & DataTable (#58)
* [CHANGE] cast income calculations to int for accurate results in BeltSurveyEntry

* [CHANGE] refactor CSV export function

* [ADD] DataTables customization for info-text and button styling

* [CHANGE] integrate DataTables buttons extension and export functionality
[CHANGE] added aa-beltradar-tooltip to scan text

* [CHANGE] remove unused star_id_raw field from SolarSystemFactory

* [ADD] Expired Filter

* [FIX] wrong Path

* [ADD] implement tests for price and income calculations in BeltSurveyEntry

* [CHANGE] enhance session details display with additional finish time and ETA elements

* [CHANGE] implement session progress styling and overlay elements

* [CHANGE] session ETA display and tooltip functionality; add translation for 'Done'

+++ beltradar/models/beltradar.py
@@ -150,7 +150,7 @@ class BeltSurveyEntry(models.Model):
         Returns:
             float: The estimated income in ISK per hour based on the price.
         """
-        return self.price_per_m3 * self.session.br_entries.rate_per_s()
+        return int(self.price_per_m3 * self.session.br_entries.rate_per_s())
 
     @property
     def income_cmp_per_h(self):
@@ -163,7 +163,7 @@ class BeltSurveyEntry(models.Model):
         Returns:
             float: The estimated income in ISK per hour based on the compressed price.
         """
-        return self.price_cmp_per_m3 * self.session.br_entries.rate_per_s()
+        return int(self.price_cmp_per_m3 * self.session.br_entries.rate_per_s())
 
     @property
     def price_per_m3(self):
@@ -177,7 +177,7 @@ class BeltSurveyEntry(models.Model):
         """
         if not self.price or not self.volume_left:
             return 0
-        return self.price / self.eve_type.volume
+        return int(self.price / self.eve_type.volume)
 
     @property
     def price_cmp_per_m3(self):
@@ -192,7 +192,7 @@ class BeltSurveyEntry(models.Model):
         """
         if not self.price_compressed or not self.volume_left:
             return 0
-        return self.price_compressed / (self.eve_type.volume / 100)
+        return int(self.price_compressed / (self.eve_type.volume / 100))
 
     # pylint: disable=useless-parent-delegation
     def save(self, *args, **kwargs):
